# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for a home-appliance store: clients, products, suppliers, and product images.",
        "title": "ShopAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/images/{id}": {
            "get": {
                "description": "Returns the raw image bytes by the image's unique identifier",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Get an image by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "invalid id format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "image not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an image by its unique identifier; the owning product's image_id is set to null",
                "tags": [
                    "images"
                ],
                "summary": "Delete an image by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Image deleted successfully"
                    },
                    "400": {
                        "description": "invalid id format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "image not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Replaces the bytes of an existing image by its ID",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Replace an existing image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Image updated successfully"
                    },
                    "400": {
                        "description": "invalid id format or failed to read request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "image not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/products/{id}/image": {
            "get": {
                "description": "Returns the raw image bytes linked to the given product",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Get a product's image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "invalid id format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "product not found or has no image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Uploads raw image bytes and links them to the given product, in a single transaction",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Add an image to a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Image added successfully"
                    },
                    "400": {
                        "description": "invalid id format or failed to read request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}